Storing users' first and last names.
The program now stores the users' preferred first and last names.
To ensure that splitting the Entries remains intact, they are stored in
the file as "$firstName+$lastname".
The login screen and main display screens' display panels have been
updated to reflect this change.
AdminScreen still needs to be updated.